distinguished telegram api logic from react component to function in utils and fixed bug with sending other reviews
